+++ targets.h
@@ -74,7 +74,7 @@ public:
   /* The maximum number of entries we want to allow storing in defer_buffer. */
   static const unsigned int DEFER_LIMIT = 64;
 
-  HostGroupState(int lookahead, int randomize, int num_random, int argc, const char *argv[]);
+  HostGroupState(int lookahead, int randomize, unsigned long num_random, int argc, const char *argv[]);
   ~HostGroupState();
   Target **hostbatch;
